+++ src/simulation/commands/GameCommand.ts
@@ -13,6 +13,7 @@ export type CommandType =
   | "embark-battalion"
   | "disembark-caravan"
   | "garrison-battalion"
+  | "attack-with-ship"
   | "attack-target"
   | "assimilate-captives"
   | "reward-heir"
@@ -86,6 +87,11 @@ export interface GarrisonBattalionPayload {
   readonly buildingId: BuildingId;
 }
 
+export interface AttackWithShipPayload {
+  readonly shipId: CaravanId;
+  readonly targetId: CaravanId;
+}
+
 export interface AttackTargetPayload {
   readonly battalionId: BattalionId;
   readonly targetId: BattalionId | BuildingId | CaravanId;
@@ -119,6 +125,7 @@ export type GameCommand =
   | GameCommandBase<"embark-battalion", EmbarkBattalionPayload>
   | GameCommandBase<"disembark-caravan", DisembarkCaravanPayload>
   | GameCommandBase<"garrison-battalion", GarrisonBattalionPayload>
+  | GameCommandBase<"attack-with-ship", AttackWithShipPayload>
   | GameCommandBase<"attack-target", AttackTargetPayload>
   | GameCommandBase<"assimilate-captives", AssimilateCaptivesPayload>
   | GameCommandBase<"reward-heir", HeirFeedbackPayload>